src/test/java/edu/hw1/Task8Test.java
package edu.hw1;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import static org.assertj.core.api.Assertions.assertThat;

public class Task8Test {
    @Test
    @DisplayName("Test knightBoardCapture true")
    public void testKnightBoardCaptureTrue() {
        int[][] board = {
            {0, 0, 0, 1, 0, 0, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 1, 0, 0, 0, 1, 0, 0},
            {0, 0, 0, 0, 1, 0, 1, 0},
            {0, 1, 0, 0, 0, 1, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 1, 0, 0, 0, 0, 0, 1},
            {0, 0, 0, 0, 1, 0, 0, 0}
        };
        assertThat(Task8.knightBoardCapture(board)).isTrue();
    }

    @Test
    @DisplayName("Test knightBoardCapture false 1")
    public void testKnightBoardCaptureFalse1() {
        int[][] board = {
            {1, 1, 1, 1, 1, 1, 1, 0},
            {0, 1, 0, 1, 1, 1, 0, 1},
            {0, 0, 0, 0, 1, 0, 1, 0},
            {0, 0, 1, 0, 1, 1, 0, 1},
            {1, 0, 0, 0, 1, 0, 1, 0},
            {0, 0, 0, 0, 1, 1, 0, 1},
            {1, 0, 0, 0, 1, 0, 1, 0},
            {0, 0, 0, 1, 1, 1, 0, 1}
        };
        assertThat(Task8.knightBoardCapture(board)).isFalse();
    }

    @Test
    @DisplayName("Test knightBoardCapture false 2")
    public void testKnightBoardCaptureFalse2() {
        int[][] board = {
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0},
            {0, 0, 0, 0, 0, 0, 0, 0}
        };
        assertThat(Task8.knightBoardCapture(board)).isTrue();
    }

}
